refactor: /services/general refactored to other -helper primitives
