+++ src/services/CacheManager.mts
@@ -1,7 +1,7 @@
 import { arrayFirst, safeArray } from './primitives/array-helper.mjs'
 import { ArrayOrSingle } from '../models/types.mjs'
 import { IIdValue } from '../models/IdValueManager.mjs'
-import { isNullOrUndefined } from './general.mjs'
+import { isNullOrUndefined } from './primitives/object-helper.mjs'
 
 type CacheManagerObject<T = object> = { expire: number; obj: T }
