+++ src/services/ContextManager.mts
@@ -6,7 +6,7 @@ import {
 import { IIdRequired, IdManager } from '../models/IdManager.mjs'
 import { IName, IValue } from '../models/interfaces.mjs'
 import { IUpdated } from '../models/id-created-updated.mjs'
-import { newGuid } from './general.mjs'
+import { newGuid } from './primitives/uuid-helper.mjs'
 import { safeObject } from './primitives/object-helper.mjs'
 
 export const zIconConfiguration = z.object({
