+++ src/services/SearchRequestView.mts
@@ -13,10 +13,14 @@ import {
   safestrLowercase,
   safestrTrim,
 } from './primitives/string-helper.mjs'
-import { hasData, isNullOrUndefined, sortFunction } from './general.mjs'
+import {
+  hasData,
+  isNullOrUndefined,
+  isObject,
+  sortFunction,
+} from './primitives/object-helper.mjs'
 import { isArray, safeArray } from './primitives/array-helper.mjs'
 import { getAsNumber } from './primitives/number-helper.mjs'
-import { isObject } from './primitives/object-helper.mjs'
 
 export type ISearchRequestView = z.infer<
   typeof SearchRequestView.zSearchRequestView
