+++ src/services/index.mts
@@ -2,11 +2,10 @@ export * from './TradeClient/index.mjs'
 export * from './ApiResponseHelper.mjs'
 export * from './CacheManager.mjs'
 export * from './charting_library.mjs'
-export * from './color-helper.mjs'
+export * from './primitives/color-helper.mjs'
 export * from './ContextManager.mjs'
 export * from './CryptoHelper.mjs'
 export * from './EmitterManager.mjs'
-export * from './general.mjs'
 export * from './html-helper.mjs'
 export * from './HttpHeaderManager.mjs'
 export * from './jwt.mjs'
@@ -18,6 +17,8 @@ export * from './primitives/function-helper.mjs'
 export * from './primitives/number-helper.mjs'
 export * from './primitives/object-helper.mjs'
 export * from './primitives/string-helper.mjs'
+export * from './primitives/symbol-helper.mjs'
+export * from './primitives/uuid-helper.mjs'
 export * from './redirect-helper.mjs'
 export * from './redux-helper.mjs'
 export * from './SearchRequestView.mjs'
