+++ src/services/primitives/string-helper.mts
@@ -4,11 +4,10 @@ import {
   Typish,
 } from '../../models/types.mjs'
 import { getAsNumber, isNumber } from './number-helper.mjs'
-import { hasData, isNullOrUndefined } from '../general.mjs'
+import { hasData, isNullOrUndefined, isObject } from './object-helper.mjs'
 import { isArray, splitToArray } from './array-helper.mjs'
 import { isFunction, typishValue } from './function-helper.mjs'
 import { AppException } from '../../models/AppException.mjs'
-import { isObject } from './object-helper.mjs'
 
 export const CONST_CharsAlphabetLower = 'abcdefghijklmnopqrstuvwxyz',
   CONST_CharsAlphabetUpper = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ',
