+++ src/services/primitives/symbol-helper.mts
@@ -1,0 +1,3 @@
+export function isSymbol(value: unknown): value is symbol {
+  return typeof value === 'symbol'
+}
