+++ src/services/primitives/uuid-helper.mts
@@ -1,0 +1,13 @@
+/**
+ * Returns a new global unique identifier (GUID).
+ * @returns A global unique identifier as a 16 character string.
+ */
+export function newGuid() {
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/gu, (c) => {
+    // eslint-disable-next-line no-bitwise
+    const r = (Math.random() * 16) | 0,
+      // eslint-disable-next-line no-bitwise
+      v = c === 'x' ? r : (r & 0x3) | 0x8
+    return v.toString(16)
+  })
+}
