+++ src/services/primitives/uuid-helper.test.mts
@@ -1,0 +1,7 @@
+import { newGuid } from './uuid-helper.mjs'
+
+test('newGuid', () => {
+  const newg = newGuid()
+
+  expect(newg.length).toBe(36)
+})
