+++ src/tplot/ConfigManager.mts
@@ -5,6 +5,10 @@ import {
   UserConfigNames,
   UserConfigTypes,
 } from '../models/UserInfo.mjs'
+import {
+  deepCloneJson,
+  hasData,
+} from '../services/primitives/object-helper.mjs'
 import { AppException } from '../models/AppException.mjs'
 import type { IConfigCharts } from './ConfigCharts.mjs'
 import type { IConfigHeaderTickerBars } from './ConfigHeaderTickerBars.mjs'
@@ -15,8 +19,6 @@ import { IDashboardSetting } from './DashboardSetting.mjs'
 import { IIdValRequired } from '../models/id-val.mjs'
 import { IKeyValueShort } from '../models/key-val.mjs'
 import { IdName } from '../models/id-name.mjs'
-import { deepCloneJson } from '../services/primitives/object-helper.mjs'
-import { hasData } from '../services/general.mjs'
 import { safestrTrim } from '../services/primitives/string-helper.mjs'
 
 export type CryptoIdeasTabNames = 'crypto' | 'nft' | 'spac'
